+++ src/promptflow-evals/promptflow/evals/evaluators/_content_safety/_content_safety_base.py
@@ -4,8 +4,8 @@
 
 from abc import ABC
 
-from promptflow.evals.common.constants import EvaluationMetrics
-from promptflow.evals.common.rai_service import evaluate_with_rai_service
+from promptflow.evals._common.constants import EvaluationMetrics
+from promptflow.evals._common.rai_service import evaluate_with_rai_service
 
 
 class ContentSafetyEvaluatorBase(ABC):
